+++ scout/server/blueprints/variants/views.py
@@ -16,7 +16,6 @@ from flask import (
     redirect,
     request,
     send_file,
-    session,
     url_for,
 )
 from flask_login import current_user
@@ -53,6 +52,7 @@ def variants(institute_id, case_name):
     category = "snv"
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
     variant_type = request.args.get("variant_type", "clinical")
+    variants_stats = store.case_variants_count(case_obj["_id"], institute_id, False)
 
     if request.form.get("hpo_clinical_filter"):
         case_obj["hpo_clinical_filter"] = True
@@ -162,9 +162,6 @@ def variants(institute_id, case_name):
     variants_query = store.variants(case_obj["_id"], query=form.data, category=category)
     result_size = store.count_variants(case_obj["_id"], form.data, None, category)
 
-    # Setup variant count session with variant count by category
-    controllers.variant_count_session(store, institute_id, case_obj["_id"], variant_type, category)
-    session["filtered_variants"] = result_size
     if request.form.get("export"):
         return controllers.download_variants(store, case_obj, variants_query)
 
@@ -180,6 +177,8 @@ def variants(institute_id, case_name):
         cytobands=cytobands,
         page=page,
         expand_search=str(request.method == "POST"),
+        result_size=result_size,
+        total_variants=variants_stats.get(variant_type, {}).get(category, "NA"),
         **data,
     )
 
@@ -232,9 +231,10 @@ def sv_variants(institute_id, case_name):
     page = int(request.form.get("page", 1))
     variant_type = request.args.get("variant_type", "clinical")
     category = "sv"
-
     # Define case and institute objects
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
+    variants_stats = store.case_variants_count(case_obj["_id"], institute_id, False)
+
     if request.form.get("hpo_clinical_filter"):
         case_obj["hpo_clinical_filter"] = True
 
@@ -255,11 +255,8 @@ def sv_variants(institute_id, case_name):
 
     variants_query = store.variants(case_obj["_id"], category=category, query=form.data)
 
-    # Setup variant count session with variant count by category
-    controllers.variant_count_session(store, institute_id, case_obj["_id"], variant_type, category)
     result_size = store.count_variants(case_obj["_id"], form.data, None, category)
 
-    session["filtered_variants"] = result_size
     # if variants should be exported
     if request.form.get("export"):
         return controllers.download_variants(store, case_obj, variants_query)
@@ -279,6 +276,8 @@ def sv_variants(institute_id, case_name):
         manual_rank_options=MANUAL_RANK_OPTIONS,
         page=page,
         expand_search=str(request.method == "POST"),
+        result_size=result_size,
+        total_variants=variants_stats.get(variant_type, {}).get(category, "NA"),
         **data,
     )
 
@@ -288,8 +287,9 @@ def sv_variants(institute_id, case_name):
 def cancer_variants(institute_id, case_name):
     """Show cancer variants overview."""
     category = "cancer"
-
+    variant_type = request.args.get("variant_type", "clinical")
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
+    variants_stats = store.case_variants_count(case_obj["_id"], institute_id, False)
 
     user_obj = store.user(current_user.email)
     if request.method == "POST":
@@ -345,7 +345,6 @@ def cancer_variants(institute_id, case_name):
 
     cytobands = store.cytoband_by_chrom(case_obj.get("genome_build"))
 
-    variant_type = request.args.get("variant_type", "clinical")
     variants_query = store.variants(case_obj["_id"], category="cancer", query=form.data)
     result_size = store.count_variants(case_obj["_id"], form.data, None, category)
 
@@ -364,6 +363,8 @@ def cancer_variants(institute_id, case_name):
             **CANCER_SPECIFIC_VARIANT_DISMISS_OPTIONS,
         },
         expand_search=str(request.method == "POST"),
+        result_size=result_size,
+        total_variants=variants_stats.get(variant_type, {}).get(category, "NA"),
         **data,
     )
 
@@ -376,9 +377,10 @@ def cancer_sv_variants(institute_id, case_name):
     page = int(request.form.get("page", 1))
     variant_type = request.args.get("variant_type", "clinical")
     category = "cancer_sv"
-
     # Define case and institute objects
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
+    variants_stats = store.case_variants_count(case_obj["_id"], institute_id, False)
+
     if request.form.get("hpo_clinical_filter"):
         case_obj["hpo_clinical_filter"] = True
 
@@ -388,11 +390,8 @@ def cancer_sv_variants(institute_id, case_name):
     cytobands = store.cytoband_by_chrom(case_obj.get("genome_build"))
     variants_query = store.variants(case_obj["_id"], category=category, query=form.data)
 
-    # Setup variant count session with variant count by category
-    controllers.variant_count_session(store, institute_id, case_obj["_id"], variant_type, category)
     result_size = store.count_variants(case_obj["_id"], form.data, None, category)
 
-    session["filtered_variants"] = result_size
     # if variants should be exported
     if request.form.get("export"):
         return controllers.download_variants(store, case_obj, variants_query)
@@ -416,6 +415,8 @@ def cancer_sv_variants(institute_id, case_name):
         cytobands=cytobands,
         page=page,
         expand_search=str(request.method == "POST"),
+        result_size=result_size,
+        total_variants=variants_stats.get(variant_type, {}).get(category, "NA"),
         **data,
     )